chore: move app usage related functions to the same file
